define gaussian_unitary_circuit
